Add legacy shwitch for -sbox-call
From: Toni Timonen

The --sbox-call means that the name of the binary and every argument,
including argv[0], is given in command line. In some cases this can't
really be resolved from the actual binary called, so it is being
preserved. This is used from -misc-runner.

Toni:

"If I remember correctly, the first argument after --sbox-call (that is
the actual binary name) is used to locate the actual binary that is
being called, but rest of the arguments are given to the elf (eg. we
skipped the program name, but kept the argv[0]; in scratchbox these
two can be different in quite many cases)."