Add elementary OMAP3530 Beagleboard support
Merged code from http://code.google.com/p/qemu-omap3/ and patched some small things here and there. There is a lot of things to check and do to get the OMAP3 support to an acceptable level. This commit provides the bare minimum level to get OMAP3 Linux running on the emulated hardware.