implement CLONE_CHILD_CLEARTID
LauroV:

I know that the clone implementation is far from the perfection and it
is very hard to implement all the clone flags using a high level API
(pthread).

This patch doesn't break seriously the host libc/libpthread. Pthread
uses the tid_address only for pthread_join. So, actually, this patch
breaks the host pthread_join (and pthread_timedjoin_np), but it makes
the emulated pthread_join work. As the qemu doesn't use pthread_join, I
think it worth to apply this patch.

Riku:

Without this patch, even the simplest threaded apps fail to run.