fix OMAP3 MMC read operation to FIFO
Reading from MMC/SD card to the OMAP3 MMC controller's FIFO buffer expected the FIFO buffer to be cleared - thus if old data was laying around in the FIFO buffer next read would yield incorrect data. However, there was a clear operation implemented in the MMC controller's data register READ operation, i.e. effectively clearing out the FIFO contents as it is read out, this problem did not surface until something was written to the MMC as the data written to the MMC stays in the FIFO after the WRITE operation causing the next READ operation after the WRITE operation to return invalid data. Sigh.