Don't globally disable kvm if one target doesn't support it
When iterating through each element in target_list, we disable kvm if we find
a target that doesn't support kvm.  This means that kvm can get globally
disabled when configuring with multiple targets.

Instead, use a new variable, has_kvm, to indicate whether the target has kvm
support or not.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>