use qemu_malloc and friends consistently
qemu_malloc, qemu_free and friends are not used consistently in the qemu
source code.

This is an attempt to use these oveloaded functions consistently all over
the place instead of the default glibc versions.

version 2 avoids to use qemu_malloc in qemu-io.c if the size to allocate is 0.

Since the submission of version 2 of the patch a lot of places have been
patched. But there are still a lot of other places where malloc/free is used
instead of qemu_malloc/qemu_free.

Version 3 of the patch is applying the fix where it has not yet been applied.

Signed-off-by: Jean-Christophe DUBOIS <jcd@tribudubois.net>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>