Set slots more carefully
KVM only supports slot sizes of PAGE_SIZE granilarity. On PPC the OS
sets the framebuffer to some odd size though, causing the current code
to simply abort().

So let's bet graceful here. We can just allocate memory sizes that are of
PAGE_SIZE granularity and everything that exceeds that comes in as MMIO and
gets handled too - just slower.

Signed-off-by: Alexander Graf <agraf@suse.de>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>