Introduce macro for defining qdev properties
This patch introduces macros for defining qdev properties.  The default macro
is clever enough to infer the type of the structure field and to automatically
generate a name for the property.  Additional macros are provided that allow
infered values to be overridden along with a set of macros to define properties
with default values.

This patch converts over syborg and a few other devices as a demonstration.
Final patch will convert everything.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>