Isolate property variables in structures
qdev properties are portions of a device's state that are writable before
construction and then readable for the remainder of the device's life cycle.
This patch introduces a best practice which strengthens that relationship.

It isolates the qdev properties in the device state with an appropriate comment.
It also moves the property list definition to follow the device state definition
in order to encourage consistency between the two.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>