Allow a cross compiler to be used for rom builds
The host tool chain is not enough to build all of the possible roms for qemu.
This patch introduces a new option, --i386-cross-prefix, that provides the
cross prefix for a toolchain to build i386 roms.

This converts multiboot to use that toolchain.  If you don't specify
a cross prefix, then the host tool chain will be used if the architecture
matches.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>