Introduce QEMU dictionary data type
A dictionary is a high-level data type used to store a collection of
values, where a unique key is associated with one value. Usually, the
notation 'key:value' is used to denote this relationship.

This commit adds the qdict module, which implements a dictionary for
QEMU and exports the following functions:

- qdict_new()     Create a new dictionary
- qdict_add()     Add a new 'key:value' pair
- qdict_get()     Get the 'value' of a given 'key'
- qdict_exists()  Check if a given 'key' exists
- qdict_del()     Delete a 'key:value' pair
- qdict_destroy() Destroy the dictionary
- qdict_size()    Return the size of the dictionary

qdict module uses a chained hash table, whose hash function comes from
module-init-tools program.

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>