monitor: Setup a dictionary with handler arguments
With this commit monitor_handle_command() will be able to setup a
QEMU dictionary with arguments to command handlers.

However, the current 'args' method is still being used, next changes
will port commands to use the new dictionary.

There are three changes introduced by the dictionary that are worth
noting:

1. The '/' argument type always adds the following standard keys in the
dictionary: 'count', 'format' and 'size'. This way, the argument
name used in the 'args_type' string doesn't matter

2. The 'l' argument type always adds a 'high' order value and a 'low'
order value. To do this the Monitor will append '_h' and '_l' to the
argument name used in the 'args_type' (this is the job of
key_append_str() & friends)

3. The optional argument type '?' doesn't need to pass the additional
'has_arg' argument, instead hanlders can do the same check with
qdict_exists()

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>