monitor: New GET_TLONG and GET_TPHYSADDR macros
When we start porting handlers to use the Monitor's dictionary
to pass argments, we will turn function parameters into automatic
variables.

This will make the build brake when the 32 bits versions of
GET_TLONG and GET_TPHYSADDR are used, because the 'h' argument
will not be used.

The best solution I could think for this problem was changing
both macros to reassign the 'h' parameter when compiled for
32 bits.

I'm open for better solutions, though.

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>