fix tsc2005 emulation to work with n900
this is more of a hack; the linux kernel module for tsc2005 will check
that the pressure value calculated from the x, z1, z2 and x_ohm does not
exceed a limit (1500). with the hardcoded settings in the tsc2005
emulation, this limit is easily exceeded. by adjusting the values we
get into the acceptable zone. however this would break the n810 emulation
which apparently uses other values or does not check the limit at all.
so instead of replacing the hardcoded defaults, make them adjustable
and use different values for n810 and n900.

Signed-off-by: Juha Riihimäki <juha.riihimaki@nokia.com>