Add json->qobject parser
This implements a json parser that generates qobjects.  The intention is that
this all can be eventually moved to a shared library to form the basis of the
QMP client API.  Unfortunately, qobject is currently GPL licensed so hopefully
Luiz will switch it to LGPL.

The parser is a pretty straight forward recursive decent parser.  It's based
on the grammar available at json.org.  The only thing it doesn't implement
is support for the 'null' keyword.  We have to decide how we want to implement
it.

It also introduces an extension to json--the ability to support strings quoted
with a single quote (similar to python).  This makes writing dictionaries
considerably nicer in C.

This module also introduces a varargs parsing function that allows a printf-like
format to be used with the parsing function.

Finally, the parser returns the amount of input consumed.  This is to make it
possible to determine the message boundaries in a stream of input.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>