Add unit test for json parser
I've attempted to make this a very thorough unit test for the json parser.  It
covers every rule and tries to test the corner cases of each rule.

There's some interesting things in this test case like a literal qobject
type.  It may be worth extract that into common code.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>