Add support for -net tap,helper=foo
Today, a user has two options for tap-based networking.  The first is to open
a tap device in a "launcher" program and pass it during exec().

The second is for qemu to create a tap device on it's own and to call a script
to setup the device.  This practically requires root privileges for qemu and
results in very poor security.

A better thing to do is to have qemu rely on a separate, more privileged tool
to create the file descriptor and then to pass that file descriptor via
a domain socket and SCM_RIGHTS.

This patch introduces a mechanism to do this that relies on creating a
socketpair() and running a helper program.  It assumes the helper program
carries more privileges either via suid or fscaps the later being the preferred
mechanism.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>