net tap helpers: rootless tap support for qemu (v2)
This series introduces a new way to do tap networking and then introduces a
new network backend that takes advantage of this new mechanism.

The ultimate goal of this series is to allow rootless bridge support in qemu.

The first patch introduces an extension to -net tap that allows for the tap fd
to come from an exec()'d helper that uses a pretty simple protocol.

The following second patches introduce a helper, qemu-bridge-helper, that allows
an unprivileged user to add a tap device to a bridge.

The third and four patch improve the security of this helper by introducing an
ACL mechanism and by using libcap to reduce privileges when invoked as SUID.

The fifth patch introduces an issue mechanism to improve the user experience
when a bridge doesn't exist.

The sixth patch enables cap_net_admin or suid on install.

The seventh and final patch introduces a -net bridge backend that is just a
syntatic wrapper around -net tap,helper=.  As third parties introduce additional
helpers (like a netcf helper), I'd like to use the same mechanism to support
those network types.

Since v1:
 - Add support for an issue when a bridge is not available
 - Various fixes from review comments
 - Improved error reporting
 -