Introduce -net bridge
This patch introduces a simple syntatic wrapper that expands from:

  -net bridge[,br=dev]

To:

  -net tap,helper="%{libexec}/qemu-bridge-helper --bridge=dev"

It's included as both an example of how to create network helpers and to
improve the out of the box experience for people who use qemu directly from
the command line.

If dev is not specified, qemubr0 is used as the default.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>