virtio-console: qdev conversion, new virtio-serial-bus
RH-Author: Amit Shah <amit.shah@redhat.com>
Message-id: <1263461886-32077-3-git-send-email-amit.shah@redhat.com>
Patchwork-id: 6348
O-Subject: [RHEL 6 PATCH v4 2/8] virtio-console: qdev conversion,
	new virtio-serial-bus
Bugzilla: 543825
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>

This commit converts the virtio-console device to create a new
virtio-serial bus that can host console and generic serial ports. The
file hosting this code is now called virtio-serial-bus.c.

The virtio console is now a very simple qdev device that sits on the
virtio-serial-bus and communicates between the bus and qemu's chardevs.

This commit also includes a few changes to the virtio backing code for
pci and s390 to spawn the virtio-serial bus.

As a result of the qdev conversion, we get rid of a lot of legacy code.
The old-style way of instantiating a virtio console using

    -virtioconsole ...

is maintained, but the new, preferred way is to use

    -device virtio-serial -device virtconsole,chardev=...

With this commit, multiple devices as well as multiple ports with a
single device can be supported.

For multiple ports support, each port gets an IO vq pair. Since the
guest needs to know in advance how many vqs a particular device will
need, we have to set this number as a property of the virtio-serial
device and also as a config option.

In addition, we also spawn a pair of control IO vqs. This is an internal
channel meant for guest-host communication for things like port
open/close, sending port properties over to the guest, etc.

This commit is a part of a series of other commits to get the full
implementation of multiport support. Future commits will add other
support as well as ride on the savevm version that we bump up here.

Signed-off-by: Amit Shah <amit.shah@redhat.com>
---
 Makefile.target        |    2 +-
 hw/pc.c                |   11 +-
 hw/ppc440_bamboo.c     |    7 -
 hw/qdev.c              |   10 +-
 hw/s390-virtio-bus.c   |   17 +-
 hw/s390-virtio-bus.h   |    2 +
 hw/s390-virtio.c       |    8 -
 hw/virtio-console.c    |  146 -------------
 hw/virtio-console.h    |   19 --
 hw/virtio-pci.c        |   13 +-
 hw/virtio-serial-bus.c |  537 ++++++++++++++++++++++++++++++++++++++++++++++++
 hw/virtio-serial.c     |  107 ++++++++++
 hw/virtio-serial.h     |  162 +++++++++++++++
 hw/virtio.h            |    2 +-
 qemu-options.hx        |    4 +
 sysemu.h               |    6 -
 vl.c                   |   17 ++-
 17 files changed, 852 insertions(+), 218 deletions(-)
 delete mode 100644 hw/virtio-console.c
 delete mode 100644 hw/virtio-console.h
 create mode 100644 hw/virtio-serial-bus.c
 create mode 100644 hw/virtio-serial.c
 create mode 100644 hw/virtio-serial.h

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>