virtio-serial-bus: Add support for buffering guest output, throttling guests
RH-Author: Amit Shah <amit.shah@redhat.com>
Message-id: <1263461886-32077-6-git-send-email-amit.shah@redhat.com>
Patchwork-id: 6355
O-Subject: [RHEL 6 PATCH v4 5/8] virtio-serial-bus: Add support for buffering
	guest output, throttling guests
Bugzilla: 543825
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>

We have to buffer data from guest as ports may not consume all the data
in one go or the guest could be fast in sending data and the apps may
not consume at the same rate.

We keep caching data the guest sends us till a port accepts it. However,
this could lead to an OOM condition where a rogue process on the guest
could continue pumping in data while the host continues to cache it.

We introduce a per-port byte-limit property to alleviate this condition.
When this limit is reached, we send a control message to the guest
indicating it to not send us any more data till further indication. When
the number of bytes cached go lesser than the limit specified, we tell
the guest to restart sending data.

Signed-off-by: Amit Shah <amit.shah@redhat.com>
---
 hw/virtio-serial-bus.c |  214 +++++++++++++++++++++++++++++++++++++++---------
 hw/virtio-serial.c     |    6 --
 hw/virtio-serial.h     |   30 ++++++-
 3 files changed, 202 insertions(+), 48 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>