virtio-serial-bus: Remove guest buffer caching and throttling.
RH-Author: Amit Shah <amit.shah@redhat.com>
Message-id: <1264002757-5820-2-git-send-email-amit.shah@redhat.com>
Patchwork-id: 6487
O-Subject: [RHEL6 PATCH 1/4] virtio-serial-bus: Remove guest buffer caching and
	throttling.
Bugzilla: 543825
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Markus Armbruster <armbru@redhat.com>
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>

Upstream requested we make these changes before upstream acceptance.

All the virtio-serial patches have now been accepted upstream and this
patch gets us in sync with the code there.

The reason for this change is that the buffering should be done inside
the guest. We now ensure we tell the guest the number of bytes that were
successfully consumed by a port. The guest app will then do the needful
if that number was lesser than the data that it sent out.

Signed-off-by: Amit Shah <amit.shah@redhat.com>
---
 hw/virtio-serial-bus.c |  181 ++----------------------------------------------
 hw/virtio-serial.c     |    8 ++-
 hw/virtio-serial.h     |   34 +--------
 3 files changed, 16 insertions(+), 207 deletions(-)

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>