Don't require explicit invocation of i440_memory_mapping_init
i440fx needs to store the mappings of the VGA area in order to support SMM since
SMM code lives in that region.  Currently, there's an explicit function in
i440fx that is invoked at start up.  This patch changes it so that this function
is called right before this mapping is needed (IOW, before SMM is activated).

This allows -device cirrus-vga to work as expected.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>