vnc: add support for VNC_AUTH_SESSION
This allows a user to select an session during the authentication phase.  After
selecting a session, we will reauthenticate, this time with a real
authentication.

The one bit that's unfortunate in this extension is that the session selection
is done before authentication.  Since the alternative is really selection by
port, it's really not a bad option.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>