pci: byte swap as PCI interface layer
PCI devices have a fixed byte order that is independent of the host byte order.
For the most part, PCI devices are little endian.  Because we have allowed PCI
devices in the past to directly register memory functions with the CPU, these
devices have been subject to automagic swapping which means that they end up
returning target endianness instead of the fixed endianness that the device
supports.

This patch fixes this by allow devices to return native words in their interface
and allowing the PCI bus to do appropriate swapping before the data makes it
to the CPU.

Right now, most devices do this wrong so this should, in theory, fix a lot of
PCI devices on big endian targets.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>