virtio-9p: convert wstat to a state machine
When we introduce a thread pool for all file system operations, we will need
be able to execute functions asynchronously.  For now, modify the existing
code so that all synchronous file system operations are executed in the tail
call position followed by an explicit invocation of the next state in the
state machine.  This will allow us to easily convert over to purely asynchronous
handling in the near term future without a tremendous amount of code churn.

N.B. for the purposes of clarity, I've used forward declarations of statics
to make the diff as readable as possible.  The next commit will reorder the
statics to remove the need for forward declarations.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>