QMP: Introduce the qmp_capabilities command
RH-Author: Luiz Capitulino <lcapitulino@redhat.com>
Message-id: <1265916626-29934-3-git-send-email-lcapitulino@redhat.com>
Patchwork-id: 7054
O-Subject: [PATCH RHEL6 qemu-kvm 2/4] QMP: Introduce the qmp_capabilities command
Bugzilla: 557930
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>
RH-Acked-by: Daniel P. Berrange <berrange@redhat.com>

This command will be used to enable QMP capabilities advertised
by the capabilities array.

Note that it will be mandatory to issue this command in order
to make QMP functional (although this behavior is not being
enforced by this commit).

Also, as we don't have any capabilities yet, the new command
doesn't accept any arguments. I will postpone the decision for
a format for this until we get our first capability.

Finally, this command is visible from the user Monitor too, in
the meaning that you can execute it but it won't do anything.
Making it only visible in QMP is beyond this series' goal, as
it requires changes in unrelated places.

Signed-off-by: Luiz Capitulino <lcapitulino@redhat.com>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
(cherry picked from commit 4a7e1190dfadf1154721125af8510347f8a3c56b)
---
 monitor.c       |   11 +++++++++++
 qemu-monitor.hx |   14 ++++++++++++++
 2 files changed, 25 insertions(+), 0 deletions(-)

Signed-off-by: Glauber Costa <glommer@redhat.com>