net: add API to disable/enable polling
RH-Author: Michael S. Tsirkin <mst@redhat.com>
Message-id: <20100215115603.GB10581@redhat.com>
Patchwork-id: 7105
O-Subject: [PATCHv2 RHEL6.0 01/22] net: add API to disable/enable polling
Bugzilla: 562958
RH-Acked-by: Gerd Hoffmann <kraxel@redhat.com>
RH-Acked-by: Juan Quintela <quintela@redhat.com>
RH-Acked-by: Amit Shah <amit.shah@redhat.com>

When vhost is bound to a backend device, we need to stop polling it when
vhost is started, and restart polling when vhost is stopped.
Add an API for that for use by vhost, and implement in tap backend.

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
(cherry picked from commit ceb696159d569db5b2a7659ce38752398c236742)
---
 net.h     |    3 +++
 net/tap.c |    8 ++++++++
 2 files changed, 11 insertions(+), 0 deletions(-)

Signed-off-by: Glauber Costa <glommer@redhat.com>